remove legacy errors
